net/mlx5e: Restore encap rules on netdev UP event
In multipath a flow is represented by two rules, one on each of the
e-switches. When one of the ports go down we try to keep the rules intact
but in the case of encap rules, after some time the encap header becomes
invalid. In such case we go over all the encap rules of the live port
and re-add them to the port that just woke up.

Issue: 1159068
Change-Id: Ia8d8c1ff29955ba2f08d6836916655358b26a6b3
Signed-off-by: Shahar Klein <shahark@mellanox.com>
Signed-off-by: Roi Dayan <roid@mellanox.com>