net/mlx5e: Add support for offloading tc priority and chaining
We used to ignore tc priority and chains when offloading tc flows,
this patch adds the support for it. To get to a new chain,
use the tc goto action.

Tables are created on demand for each [chain,prio] pair.

Currently we support a fixed prio range 1-16, and chain range 0-3,
giving us 48 prios to play with.

Issue: 1323489
Change-Id: Ie31d40f5e62dbbb5b69f62465d3c8212134a4f96
Signed-off-by: Paul Blakey <paulb@mellanox.com>